Fixed pylint errors.
